[#12833] - Cleanup for tests
