--- tests/integration/Session/Bag/GetDICest.php
@@ -1,45 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalconphp.com>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-namespace Phalcon\Test\Integration\Session\Bag;
-
-use IntegrationTester;
-use Phalcon\Di\FactoryDefault;
-use Phalcon\Session\Bag;
-use Phalcon\Test\Fixtures\Traits\DiTrait;
-use Phalcon\Test\Fixtures\Traits\SessionBagTrait;
-
-/**
- * Class GetDICest
- */
-class GetDICest
-{
-    use DiTrait;
-    use SessionBagTrait;
-
-    /**
-     * Tests Phalcon\Session\Bag :: getDI()
-     *
-     * @param IntegrationTester $I
-     *
-     * @author Phalcon Team <team@phalconphp.com>
-     * @since  2018-11-13
-     */
-    public function sessionBagGetDI(IntegrationTester $I)
-    {
-        $I->wantToTest("Session\Bag - getDI()");
-        $session = new Bag("DiTest");
-        $di      = new FactoryDefault();
-        $session->setDI($di);
-        $I->assertEquals($di, $session->getDI());
-    }
-}
