+++ tests/integration/Session/Bag/OffsetGetSetCest.php
@@ -18,29 +18,29 @@ use Phalcon\Test\Fixtures\Traits\DiTrait;
 use Phalcon\Test\Fixtures\Traits\SessionBagTrait;
 
 /**
- * Class SetCest
+ * Class OffsetGetSetCest
  */
-class SetCest
+class OffsetGetSetCest
 {
     use DiTrait;
     use SessionBagTrait;
 
     /**
-     * Tests Phalcon\Session\Bag :: set()
+     * Tests Phalcon\Session\Bag :: offsetGet()/offsetSet()
      *
      * @param IntegrationTester $I
      *
      * @author Phalcon Team <team@phalconphp.com>
      * @since  2018-11-13
      */
-    public function sessionBagSet(IntegrationTester $I)
+    public function sessionBagOffsetGet(IntegrationTester $I)
     {
-        $I->wantToTest("Session\Bag - set()");
+        $I->wantToTest("Session\Bag - offsetGet()/offsetSet()");
         $session = new Bag("SetTest");
 
         $testValue = "TestValue";
-        $session->set("test", $testValue);
+        $session->offsetSet("test", $testValue);
 
-        $I->assertEquals($testValue, $session->get("test"));
+        $I->assertEquals($testValue, $session->offsetGet("test"));
     }
 }
